overview_tab

Provides an overview table for the time and scope conditions of a data set. If a data.table object is provided, the function uses data.table's syntax to perform the evaluation

Usage

overview_tab(
  dat,
  id,
  time = list(year = NULL, month = NULL, day = NULL),
  complex_date = FALSE
)

Arguments

dat

A data frame or data table object

id

Scope (e.g., country codes or individual IDs)

time

Time (e.g., time periods given by years, months, ...). There are three options to add a date variable: 1) Time can be a character vector containing **one** time variable, 2) a time variable following the YYYY-MM-DD format, or 3) or a list containing multiple time variables (`time = list(year = NULL, month = NULL, day = NULL)`).

complex_date

Boolean argument identifying if there is a more complex (list-wise) date_time parameter (FALSE is the default)

Value

A data frame object that contains a summary of a sample that can later be converted to a 'LaTeX' output using overview_print

Examples

# With version 1 (and also 2):

data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)

# With version 3:
overview_tab(dat = toydata, id = ccode, time = list(year = toydata$year,
             month = toydata$month, day = toydata$day), complex_date = TRUE)
#> # A tibble: 5 × 2
#> # Groups:   ccode [5]
#>   ccode time_frame                                                              
#>   <chr> <chr>                                                                   
#> 1 AGO   7305, 7337, 7366, 7398, 7429, 7461, 7492, 7524, 7556, 7587, 7619, 7650,…
#> 2 BEN   9131, 9163, 9192, 9224, 9255, 9287, 9318, 9350, 9382, 9413, 9445, 9476,…
#> 3 FRA   8401, 8433, 8462, 8494, 8525, 8557, 8588, 8620, 8652, 8683, 8715, 8746,…
#> 4 GBR   7670, 7702, 7731, 7763, 7794, 7826, 7857, 7889, 7921, 7952, 7984, 8015,…
#> 5 RWA   c(7305, 7306, 7307, 7308, 7309, 7310, 7311, 7312, 7313, 7314, 7315, 731…